Fix the two remaining lint warnings: redundant cast, this-escape in UI constructor
The UI class is now final, so registering listeners on this inside the
constructor cannot reach a partially initialised subclass.

+++ src/main/java/mangapagessplitter/BatchProcessor.java
@@ -448,7 +448,7 @@ public class BatchProcessor {
             // Update progress (for image processing within the current folder)
             if (totalImages > 0) {
                 updateProgress("Processing image " + (i + 1) + "/" + totalImages +
-                               " in " + folder.getFileName(), (int)((i * 100) / totalImages));
+                               " in " + folder.getFileName(), (i * 100) / totalImages);
             }
 
             Path imagePath = imagePaths.get(i);
